+++ raster/src/main/scala/geotrellis/raster/io/geotiff/compression/Decompressor.scala
@@ -117,6 +117,8 @@ object Decompressor {
         checkEndian(PackBitsDecompressor(segmentSizes))
       case JpegCoded =>
         checkEndian(JpegDecompressor(tiffTags))
+      case ZStdCoded =>
+        checkPredictor(ZStdCompression.createDecompressor)
 
       // Unsupported compression types
       case HuffmanCoded =>